TST add more tests
