+++ benchopt/tests/utils/capture_run_output.py
@@ -36,7 +36,7 @@ class CaptureRunOutput(object):
 
         # Make sure to delete all the result that created by the run command.
         self.result_files = re.findall(
-            r'Saving result in: (.*\.parquet)', self.output
+            r'Saving result in: (.*\.parquet|.*\.csv)', self.output
         )
         if len(self.result_files) >= 1 and self.delete_result_files:
             for result_file in self.result_files:
